Refactor AWS Lambda update functionality
The App now auto-generates the AWS Lambda function name and AWS API
Gateway name from the user-provided App name. As a result, the backend
sam/awslambda.py had to be refactored with this new support in such that
the function names would be read from the settings object rather than
requiring a function name to be provided. This lays the groundwork for a
future commit to enable user to update the AWS Lambda function code and
handler from the CLI.