Fix to use deployed function name during update
There is a distinction between the app.function_name and
app.cloud.function_name, where one is the user-defined function name,
the logical function name is defined within the cloud object. In
addition, the unit tests were fixed to remove the extraneous
'status_code' and 'data' key entries from the boto3 response. The
application only has access to the 'data' key value so the status code
is based on the 'ResponseMetadata' and 'HTTPStatusCode' key value.